@@ -4,12 +4,12 @@ plugins {
 
 android {
     namespace 'com.example.audioplayer'
-    compileSdk 34
+    compileSdk 35
 
     defaultConfig {
         applicationId "com.example.audioplayer"
         minSdk 24
-        targetSdk 34
+        targetSdk 35
         versionCode 1
         versionName "1.0"
 
@@ -18,7 +18,8 @@ android {
 
     buildTypes {
         release {
-            minifyEnabled false
+            minifyEnabled true
+            shrinkResources = true
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
@@ -29,19 +30,10 @@ android {
 }
 
 dependencies {
-
     implementation libs.appcompat
     implementation libs.material
-    implementation libs.activity
-    implementation libs.constraintlayout
-    testImplementation libs.junit
-    androidTestImplementation libs.ext.junit
-    androidTestImplementation libs.espresso.core
-
-
     implementation libs.glide
     annotationProcessor libs.compiler
     implementation libs.circleimageview
-
-
+    implementation libs.datamanager
 }
